impl header notification settings
