+++ src/components/Main/Modal/NotificationModal/NotificationStateSelector.vue
@@ -12,40 +12,11 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, computed, reactive } from '@vue/composition-api'
+import { defineComponent } from '@vue/composition-api'
 import store from '@/store'
-import { makeStyles } from '@/lib/styles'
+import useChannelSubscriptionState from '@/use/channelSubscriptionState'
 import NotificationStateSelectorItem from './NotificationStateSelectorItem.vue'
 import { SubscriptionLevel } from '@/store/domain/me'
-
-const useChannelSubscriptionState = () => {
-  const currentChannelId = computed(
-    () => store.state.domain.messagesView.currentChannelId
-  )
-  const currentChannelSubscription = computed(
-    () =>
-      store.state.domain.me.subscriptionMap[currentChannelId.value] ?? 'none'
-  )
-  const changeSubscriptionLevel = (level: SubscriptionLevel) => {
-    store.dispatch.domain.me.changeSubscriptionLevel({
-      channelId: currentChannelId.value,
-      subscriptionLevel: level
-    })
-  }
-  const changeNextSubscriptionLevel = () => {
-    const level =
-      currentChannelSubscription.value === 'notified'
-        ? 'none'
-        : currentChannelSubscription.value === 'subscribed'
-        ? 'notified'
-        : 'subscribed'
-  }
-  return {
-    currentChannelSubscription,
-    changeSubscriptionLevel,
-    changeNextSubscriptionLevel
-  }
-}
 
 const levels: SubscriptionLevel[] = ['notified', 'subscribed', 'none']
